+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.9.0
+
+- New `ShadContextMenu` component.
+- Add `groupId` to `ShadPopover`, to determine if the tap is inside the popover or not.
+- Add `onFocusChange` to `ShadFocusable` and `ShadButton`.
+
 ## 0.8.0
 
 - **BREAKING CHANGE**: Refactor `ShadResizablePanelGroup` in order to react to window resize correctly. The sizes have been normalized. You don't need to provide anymore a pixel size, but a value between 0 and 1 which indicates the percentage of the available space.